BJS2-66099-post-system fix mapper, sort test in the folders
