+++ src/main/java/faang/school/postservice/mapper/PostMapper.java
@@ -7,6 +7,7 @@ import faang.school.postservice.model.Comment;
 import faang.school.postservice.model.Like;
 import faang.school.postservice.model.Post;
 import faang.school.postservice.model.Resource;
+import org.mapstruct.BeanMapping;
 import org.mapstruct.Mapper;
 import org.mapstruct.Mapping;
 import org.mapstruct.MappingTarget;
@@ -16,6 +17,11 @@ import java.util.List;
 @Mapper(componentModel = "spring")
 public interface PostMapper {
 
+    @BeanMapping(ignoreByDefault = true)
+    @Mapping(target = "content", source = "content")
+    @Mapping(target = "authorId", source = "authorId")
+    @Mapping(target = "projectId", source = "projectId")
+    @Mapping(target = "scheduledAt", source = "scheduledAt")
     Post createDtoToEntity(PostCreateDto postCreateDto);
 
     @Mapping(source = "likes", target = "totalLikes")
@@ -24,6 +30,11 @@ public interface PostMapper {
     @Mapping(source = "ad.id", target = "adId")
     PostViewDto toViewDto(Post post);
 
+    @BeanMapping(ignoreByDefault = true)
+    @Mapping(target = "content", source = "content")
+    @Mapping(target = "published", source = "published")
+    @Mapping(target = "deleted", source = "deleted")
+    @Mapping(target = "scheduledAt", source = "scheduledAt")
     void update(PostUpdateDto source, @MappingTarget Post target);
 
     default long getTotalLikes(List<Like> likes) {
